troop and enemy stuff

+++ CoKe/Comparable.pde
@@ -1,5 +1,5 @@
 interface Comparable<T>{
   
-  int compareTo(T other);
+  float compareTo(T other);
   
 }
